requirements.txt
googleapis-common-protos==1.5.3
grpcio==1.14.1
grpcio-tools==1.14.1
protobuf==3.6.1
PyScaffold==3.0.3
